+++ ami-api/__init__.py
@@ -1,6 +1,23 @@
 import os
 
 from flask import Flask
+from logging.config import dictConfig
+
+dictConfig({
+    'version': 1,
+    'formatters': {'default': {
+        'format': '[%(asctime)s] %(levelname)s in %(module)s: %(message)s',
+    }},
+    'handlers': {'wsgi': {
+        'class': 'logging.StreamHandler',
+        'stream': 'ext://flask.logging.wsgi_errors_stream',
+        'formatter': 'default'
+    }},
+    'root': {
+        'level': 'INFO',
+        'handlers': ['wsgi']
+    }
+})
 
 def create_app(test_config=None):
     # create and configure the app
@@ -22,10 +39,7 @@ def create_app(test_config=None):
     except OSError:
         pass
 
-    from . import index
     from . import ami
-
-    app.register_blueprint(index.bp)
     app.register_blueprint(ami.bp)
 
     return app