+++ CHANGES.md
@@ -1,4 +1,10 @@
 0.1.0 (2020-12-03)
 ------------------
-
 * Initial release
+
+
+0.2.0 (2020-12-04)
+------------------
+* Implement the boto3 filter tags
+* Implement the flask logging
+* Add the Dockerfile